Use `GetTemplatedMethods` and fix logic for templated functions in ProxyWrappers
This deprecates `IsTemplatedConstructor` and performs the underlying purpose of the code block. Handle non-templated constructors and templated methods in the first loop and un-instantiated template methods and templated constructors in the second.
This PR also refactors loops in Dispatcher.cxx and ProxyWrapper.cxx to no longer use index based interfaces to retrieve template method info