Preserve Python errors when executing C++ function
**Corresponds to the following ROOT PR:**
https://github.com/root-project/root/pull/18163

The internal `CPyCppyy::CPPMethod::SetPyError_()` method currently
handles three different cases related to method calls:

1. When there is something wrong with the arguments or any other problem
   when setting up the function call, `CPyCppyy` itself will set a **new
   error**. Example usage in
   [VerifyArgCount_](https://github.com/root-project/root/blob/master/bindings/pyroot/cppyy/CPyCppyy/src/CPPMethod.cxx#L75)..
2. Adding context information when dealing with a **C++ exception** when
   [calling the
   function](https://github.com/root-project/root/blob/master/bindings/pyroot/cppyy/CPyCppyy/src/CPPMethod.cxx#L981).
3. Dealing with a **Python error** when executing the function. In this
   case, the current logic resets the exception to something that looks
   like the C++ exception, only mentioning the current `CPPMethod` and
   the final exception message. However, that means that all the
   traceback between the `CPPMethod` and the place where the error
   occurred is lost, although it was available.

This commit suggests to make the third case of dealing with a Python
exception a no-op. Like this, we keep the already very informative
Python traceback. There is only one drawback: the exception is not
augmented with information on the C++ method anymore. But given that in
case of a Python error, getting info on the Python code and traceback is
more important than getting info on the C++ side, this seems to me an
acceptable price to pay.

The motivation for this change is that Python callbacks passed to the
C++ side are increasingly common, and one doesn't want to be blind in
case of errors.

Also, refactor a code to match the `PyErr_Fetch` with a `PyErr_Restore`
when appropriate, as recommended by the Python C API documentation.

Closes ROOT issue https://github.com/root-project/root/issues/18146.

Simple demo based on the code example in the issue:
```python
import ROOT

def inner_func(x):
    raise ValueError("hello")

def func(x):
    return inner_func(x)

cpp_wrapper = ROOT.std.function["double(double)"](func)

cpp_wrapper(3)
```
Without this PR
```
Traceback (most recent call last):
  File "script.py", line 48, in main
    cpp_wrapper_1(3)
    ^^^^^^^^^^^^^^^^
ValueError: double std::function<double(double)>::operator()(double __args) =>
    ValueError: hello
```
With this PR:
```
Traceback (most recent call last):
  File "script.py", line 11, in <module>
    cpp_wrapper(3)
  File "script.py", line 7, in func
    return inner_func(x)
           ^^^^^^^^^^^^^
  File "script.py", line 4, in inner_func
    raise ValueError("hello")
ValueError: hello
```

Co-Authored-By: Jonas Rembser <jonas.rembser@cern.ch>