Avoid false positives in `size()` pythonization
When checking if a class has the `size()` method implemented in the
style of the STL, the attribute must be a CPyCppyy overload. Otherwise
the check gives false positives in the case where the class has a
non-function attribute that is called "size".

This problem was caught in the Python debug interpreter.